requirements.txt
absl-py==1.2.0
astunparse==1.6.3
cached-property==1.5.2
cachetools==5.2.0
certifi==2022.6.15.2
charset-normalizer==2.0.12
cloudpickle==2.2.0
cycler==0.11.0
flatbuffers==1.12
fonttools==4.37.1
gast==0.4.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.1
gym==0.23.1
gym-notices==0.0.8
h5py==3.6.0
idna==3.4
imageio==2.16.1
importlib-metadata==4.12.0
joblib==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
numpy==1.22.4
oauthlib==3.2.1
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.4
Pillow==9.2.0
pip==22.2.2
protobuf==3.19.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.1.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.3.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.2
setuptools==65.3.0
six==1.15.0
sklearn==0.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.9.0
tensorflow-macos==2.9.2
tensorflow-metal==0.5.0
termcolor==2.0.1
tifffile==2022.3.25
torchvision==0.12.0
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
wheel==0.37.1
wrapt==1.14.1
zipp==3.8.1
